Added assessments completed page. Fixed navbar logo with white background.
When a job-seeker finishes an assessment, they are now redirected to a page which tells them their assessment answers have been saved (`assessment_completed.html`). This page also gives them 2 buttons, one to return home (`main.html`) and one to return to assessments lists  (`assessmentsList.html`)

The navbar logo had a white background, even though the logo is a transparent PNG, so I commented out the part in `style.css` that added that background colour to the logo.