Fixed profile updating saving to a different collection than the users collection.
Previously, editing and saving your profile in `proifile.html` saved it to a profiles collection with the UID, but nothing reads the profiles collection, only the users collection. Now, when a user updates and saves their profile, its saved to the users collection with their UID so things can actually read them throuought the web app.