Recruiters can now view all posted assessments. as well as their responses in their `recruiterdashboard.html`
Clicking on "View Applications" in the recruiter dashboard will open another page, `applicationView.html` with some parameters passed. A dropdown is there to select users by their name. When selected, it  displays the title of the posting, the users email, and their answers to each question, as well as a list of correct answers.