Added job posting deletion functionality in `recruiterdashboard.html`
Clicking delete (in bright red. to the left of the "View Applications" button) sends an alert to the recruiters browser. asking if they're sure they want to delete the posting. If they click yes, the posting will be deleted and their page will be refreshed to fetch the updated information.